Extract chain utils to sdk-common
