+++ libs/sdk-core/src/test_utils.rs
@@ -12,7 +12,7 @@ use rand::distributions::{Alphanumeric, DistString, Standard};
 use rand::rngs::OsRng;
 use rand::{random, Rng};
 use sdk_common::grpc;
-use sdk_common::prelude::{FiatAPI, FiatCurrency, Rate};
+use sdk_common::prelude::{FiatAPI, FiatCurrency, OnchainTx, Rate, TxStatus};
 use serde_json::{json, Value};
 use tokio::sync::{mpsc, watch, Mutex};
 use tokio::time::sleep;
@@ -29,7 +29,7 @@ use crate::bitcoin::util::bip32::{ChildNumber, ExtendedPrivKey};
 use crate::bitcoin::Network;
 use crate::breez_services::{OpenChannelParams, Receiver};
 use crate::buy::BuyBitcoinApi;
-use crate::chain::{ChainService, OnchainTx, Outspend, RecommendedFees, TxStatus};
+use crate::chain::{ChainService, Outspend, RecommendedFees};
 use crate::error::{ReceivePaymentError, SdkError, SdkResult};
 use crate::invoice::{InvoiceError, InvoiceResult};
 use crate::lightning::ln::PaymentSecret;
